fix(lint): resolve require-root-testid violations and circular dep from census splits
- Add data-testid attrs (onb-checklist-group/header/done/completed, checklist-item-row, onb-checklist-item-label) to ChecklistGroup.tsx to carry forward #1532 testid migration
- Extract ChecklistItemRow to ChecklistItemRow.tsx and isChecklistGroupSettled/completedChecklistItems to checklist-item-helpers.ts to break ChecklistGroup <-> ChecklistSection mutual import cycle
- Add data-testid="targets-page" to TargetsPage root div; eslint-disable on MasterDetail empty-state branch (DetailPane renders data-testid="detail" internally)
- Regenerate eslint-alm-baseline.txt: 3 grandfathered entries removed (MasterDetail/TargetsPage fixed; ChecklistItemRow resolved)
